Move server setup to a separate file
